+++ NEWS.md
@@ -5,6 +5,7 @@
 
 #### Highlights
 * Fix broken @docTYPE package documentation as described here: https://github.com/r-lib/roxygen2/issues/1491
+* Update PlotMapOutput() and PlotMapPoints() to allow for missing SUBIDs in MapOutput or GIS files
 
 ## HYPEtools 1.5.0 (2023-08-14)
 *Enhancements and bug fixes*